Going to create a snake game.
Added a basic surface, loaded a block and created draw function to draw it and then move it.
Changed everything in OOP to make it more readable and easier to understand.
Made the bolck to move in every direction by itself, we chose direction and it moves.
Made the snalke longer and worte hoee the bolcks will move.
Also draw the apple.
Added score and collision fucntion,
Added end sceen and posibility tp replay